refactor: updated file path for homepage components/pages
