+++ src/components/pages/Homepage.tsx
@@ -17,6 +17,8 @@ const Homepage: React.FC = () => {
             </div>
           </div>
         </Link>
+
+        <div></div>
       </div>
     </div>
   );
